// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const RoleEnum = {
  ADMIN: 'ADMIN',
  STUDENT: 'STUDENT'
} as const

export type RoleEnum = (typeof RoleEnum)[keyof typeof RoleEnum]


export const StatusEnum = {
  IN_PROGRESS: 'IN_PROGRESS',
  SUBMITTED: 'SUBMITTED',
  EXPIRED: 'EXPIRED'
} as const

export type StatusEnum = (typeof StatusEnum)[keyof typeof StatusEnum]